+++ src/components/ui/Chat/ChatBubble/ChatBubbleContainer.tsx
@@ -10,49 +10,63 @@ import { Dots } from '../../icons';
 import { formatTime } from '../../Utilities';
 
 export type Message = {
-  /* The displayed text of the message sent. */
+  /** The displayed text of the message sent. */
   content: string;
-  /* The timestamp when the message was originally sent by the sender. */
+  /** The timestamp when the message was originally sent by the sender. */
   createdTimestamp: string;
-  /* The timestamp of the last time the message was edited. */
-  lastUpdatedTimestamp?: string;
-  /* Determines if a message was redacted (deleted) by a user. */
+  /** The timestamp of the last time the message was edited. */
+  lastEditedTimestamp?: string;
+  /** Determines if a message was redacted (deleted) by a user. */
   redacted: boolean;
-  /* The display name of the sender. */
+  /** The display name of the sender. */
   senderName: string;
-  /* The unique identifier of the sender. */
+  /** The unique identifier of the sender. */
   senderId: string;
 };
 
 export type MessageVariant = 'outgoing' | 'incoming';
 
-export interface ChatBubbleContainerProps extends Omit<HTMLAttributes<HTMLDivElement>, 'css'>, BaseProps {
-  /* The time the message was originally sent. */
-  createdTimestamp: string;
-  /* This allows for custom elements to be rendered into the optional popover menu for each message, such as an edit button. */
+export interface ChatBubbleContainerProps
+  extends Omit<HTMLAttributes<HTMLDivElement>, 'css'>,
+    BaseProps {
+  /** The time the message was originally sent. */
+  timestamp?: string;
+  /** This allows for custom elements to be rendered into the optional popover menu for each message, such as an edit button. */
   actions?: ReactNode | ReactNode[];
-};
+}
 
 export const ChatBubbleContainer: FC<ChatBubbleContainerProps> = (props) => {
-  const { createdTimestamp, actions } = props;
+  const { timestamp, actions } = props;
 
   return (
-    <StyledChatBubbleContainer data-testid='chat-bubble'>
+    <StyledChatBubbleContainer
+      data-testid="chat-bubble-container"
+      actions={actions}
+    >
       {props.children}
       <StyledChatBubbleInfo>
-        <span data-testid='message-time'>{formatTime(createdTimestamp)}</span>
-        {actions &&
-        <PopOver
-          a11yLabel="Open channel options"
-          placement="bottom-end"
-          renderButton={isOpen => (
-            <Dots width='2rem' height='2rem' className={`${isOpen ? 'isOpen' : ''} ch-message-actions`} />
-          )}
-          children={actions}
-        />}
+        {timestamp && (
+          <span className="ch-timestamp" data-testid="message-time">
+            {formatTime(timestamp)}
+          </span>
+        )}
+        {actions && (
+          <PopOver
+            a11yLabel="Open channel options"
+            placement="bottom-end"
+            renderButton={(isOpen) => (
+              <Dots
+                width="2rem"
+                height="2rem"
+                className={`${isOpen ? 'isOpen' : ''} ch-message-actions`}
+              />
+            )}
+            children={actions}
+          />
+        )}
       </StyledChatBubbleInfo>
     </StyledChatBubbleContainer>
-  )
+  );
 };
 
 export default ChatBubbleContainer;