+++ src/components/ui/Chat/ChatBubble/Styled.tsx
@@ -3,39 +3,50 @@
 
 import styled from 'styled-components';
 
-import { baseStyles, baseSpacing } from '../../Base';
+import { baseStyles, baseSpacing, BaseProps } from '../../Base';
 import { MessageVariant } from './ChatBubbleContainer';
 
-interface StyledChatBubbleProps {
+interface StyledChatBubbleContainerProps extends BaseProps {
+  actions?: any;
+}
+
+interface StyledChatBubbleProps extends BaseProps {
   variant: MessageVariant;
   editable?: boolean;
 }
 
-export const StyledChatBubbleContainer = styled.div`
+export const StyledChatBubbleContainer = styled.div<
+  StyledChatBubbleContainerProps
+>`
   display: flex;
   flex-direction: row;
-  margin: 0.5rem 0;
   font-size: 0.65rem;
   margin-left: 1rem;
   align-items: center;
+  background-color: ${(props) => props.theme.chatBubble.container.bgd};
+  width: 100%;
+
+  .ch-timestamp {
+    padding-right: ${(props) => (props.actions ? '' : '2rem')};
+  }
 
   ${baseSpacing}
   ${baseStyles}
 `;
 
 export const StyledChatBubble = styled.div<StyledChatBubbleProps>`
-  background-color: ${props => props.theme.chatBubble[props.variant].bgd};
+  background-color: ${(props) => props.theme.chatBubble[props.variant].bgd};
   padding: 0.625rem 1rem;
   border-radius: 4px;
   width: fit-content;
-  color: ${props => props.theme.chatBubble[props.variant].fontColor};
+  color: ${(props) => props.theme.chatBubble[props.variant].fontColor};
   box-shadow: 0 1px 4px 0 rgba(0, 0, 0, 0.1);
   line-height: 20px;
-  width: ${props => props.editable ? '100%' : 'fit-content'};
+  width: ${(props) => (props.editable ? '100%' : 'fit-content')};
   max-width: 70.6%;
   font-size: 0.875rem;
   position: relative;
-  margin-bottom: ${props => props.editable ? '4rem' : 'auto'};
+  margin-bottom: ${(props) => (props.editable ? '4rem' : 'unset')};
 
   & .ch-sender-name {
     font-weight: bold;
@@ -50,7 +61,7 @@ export const StyledChatBubble = styled.div<StyledChatBubbleProps>`
     margin-left: -20px;
 
     & .ch-chat-bubble-tail {
-      fill: ${props => props.theme.chatBubble[props.variant].bgd}
+      fill: ${(props) => props.theme.chatBubble[props.variant].bgd};
     }
   }
 
@@ -62,20 +73,20 @@ export const StyledChatBubble = styled.div<StyledChatBubbleProps>`
     position: absolute;
     margin-top: 1rem;
     display: flex;
+    flex-direction: row-reverse;
     width: 10rem;
     justify-content: space-between;
   }
 
-  .ch-edited {
-    font-style: italic;
-  }
+  ${baseSpacing};
+  ${baseStyles};
 `;
 
 export const StyledChatBubbleInfo = styled.div`
   display: flex;
   margin-right: 0.5rem;
   margin-left: auto;
-  color: ${props => props.theme.chatBubble.container.fontColor};
+  color: ${(props) => props.theme.chatBubble.container.fontColor};
   align-items: center;
 
   & .ch-message-actions {
@@ -83,28 +94,28 @@ export const StyledChatBubbleInfo = styled.div`
   }
 
   & button:hover .ch-message-actions {
-    background-color: ${props => props.theme.buttons.icon.hover.bgd};
+    background-color: ${(props) => props.theme.buttons.icon.hover.bgd};
 
     & g {
-      fill: ${props => props.theme.buttons.icon.hover.text};
+      fill: ${(props) => props.theme.buttons.icon.hover.text};
     }
   }
 
   & button:focus .ch-message-actions {
-    border: 1px solid ${props => props.theme.colors.primary.dark};
+    border: 1px solid ${(props) => props.theme.colors.primary.dark};
     border-radius: 50%;
   }
 
   & .ch-message-actions g {
-    fill: ${props => props.theme.chatBubble.container.fontColor};
+    fill: ${(props) => props.theme.chatBubble.container.fontColor};
   }
 
   & .ch-message-actions.isOpen {
-    background-color: ${props => props.theme.buttons.icon.active.bgd};
+    background-color: ${(props) => props.theme.buttons.icon.active.bgd};
     border-radius: 50%;
 
     & g {
-      fill: ${props => props.theme.buttons.icon.active.text};
+      fill: ${(props) => props.theme.buttons.icon.active.text};
     }
   }
 `;