+++ tst/components/ui/Chat/DateHeader/index.test.tsx
@@ -4,15 +4,16 @@
 import '@testing-library/jest-dom';
 import React from 'react';
 
-import DateHeader, { DateHeaderProps } from '../../../../../src/components/ui/Chat/DateHeader';
+import DateHeader, {
+  DateHeaderProps,
+} from '../../../../../src/components/ui/Chat/DateHeader';
 import lightTheme from '../../../../../src/theme/light';
 import { renderWithTheme } from '../../../../test-helpers';
 describe('DateHeader', () => {
-
   const defaultProps: DateHeaderProps = {
     date: new Date().toString(),
-    key: '1'
-  }
+    key: '1',
+  };
 
   it('should render a DateHaader component in the document', () => {
     const component = <DateHeader {...defaultProps} />;